+++ client/list.js
@@ -36,20 +36,30 @@ addHosterForm.addEventListener('submit', async (event) => {
 });
 
 async function deleteHoster(uuid) {
-	const response = await fetch(API_URL + 'hoster/delete/' + uuid, {
-		method: 'GET',
-		headers: {
-			'auth-token': getCookie('auth-token'),
-		},
+	const { isConfirmed: confirmed } = await Swal.fire({
+		title: 'Error!',
+		text: 'Do you really want to delete the Hoster?',
+		icon: 'warning',
+		showCancelButton: true,
+		cancelButtonText: 'No im stupid!',
+		confirmButtonText: 'Yes im sure!'
 	});
-	const json = await response.json();
-	if (json.success) {
-		loadHosters();
-	} else {
-		if (json.message.includes('auth-token')) {
-			deleteCookie('auth-token');
+	if (confirmed) {
+		const response = await fetch(API_URL + 'hoster/delete/' + uuid, {
+			method: 'GET',
+			headers: {
+				'auth-token': getCookie('auth-token'),
+			},
+		});
+		const json = await response.json();
+		if (json.success) {
+			loadHosters();
 		} else {
-			alert('Error: ' + json.message);
+			if (json.message.includes('auth-token')) {
+				deleteCookie('auth-token');
+			} else {
+				alert('Error: ' + json.message);
+			}
 		}
 	}
 }
